Alert if we have print sent content without newspaper book section tag.
