+++ app/com/gu/contentapi/sanity/MetaSuites.scala
@@ -11,7 +11,7 @@ object MetaSuites {
     new PreviewRequiresAuthTest(context),
     new ContentApiSanityTest(context),
     new GetNonExistentContentShould404(context),
-    new PreviewContentSetNotInLiveTest(context),
+    new ContentSetFeaturesTest(context),
     new ValidateArticleSchema(context),
     new NewestItemFieldsTest(context),
     new MostViewedContainsItemsTest(context),
